requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
flake8
python-dotenv
django-filter